Cleanup textproto
